better JSON parsing in client

+++ src/benchmarks/edu/mit/benchmark/b2w/procedures/CreateStockTransaction.java
@@ -61,7 +61,7 @@ public class CreateStockTransaction extends VoltProcedure {
             ");");
 
     public VoltTable[] run(String transaction_id, String[] reserve_id, String[] brand, TimestampType[] timestamp,
-        TimestampType[] expiration_date, int[] is_kit, int[] requested_quantity, String[] reserve_lines, int[] reserved_quantity, long[] sku, 
+        TimestampType[] expiration_date, byte[] is_kit, int[] requested_quantity, String[] reserve_lines, int[] reserved_quantity, long[] sku, 
         String[] solr_query, long[] store_id, int[] subinventory, int[] warehouse) {
         
         String current_status = B2WConstants.STATUS_NEW;
